ai/requirements.txt
mlflow==2.16.0
pandas==2.2.2
pyarrow==16.1.0
Pillow==10.4.0
python-dotenv==1.0.1
